change: move application types to applications folder
