+++ src/types/applications/application_flags.ts
