+++ src/types/mod.ts
@@ -196,8 +196,8 @@ export * from "./misc/image_format.ts";
 export * from "./misc/image_size.ts";
 export * from "./misc/typing_start.ts";
 export * from "./mod.ts";
-export * from "./oauth2/application.ts";
-export * from "./oauth2/application_flags.ts";
+export * from "./applications/application.ts";
+export * from "./applications/application_flags.ts";
 export * from "./oauth2/bot_auth_query.ts";
 export * from "./oauth2/get_current_authorization_information.ts";
 export * from "./oauth2/scopes.ts";
